move to `get_scheduler`
